fix(mux): restore into a member's own worktree cwd, not the squad's origins[0]
x-caef PR2 part 2. Confirmed unfixed in source per the operator's audit:
StoredMember carried no cwd, so restore_squads always spawned every member's
pane at cwd0 = ps.origins.first() (server.rs). A worker that lived in
<repo>/.claude/worktrees/x-caef restored into the repo root instead, silently
- the operator discovers it only when the agent starts editing the wrong
checkout.

StoredMember.cwd is additive (#[serde(default)]) and re-derived on every
persist from the live pane's own captured cwd, the same posture persist_squad
already uses for tab_name. Restore spawns at the stored cwd when it still
resolves (case 2: unchanged identity, nothing to notice) and falls back to
origins[0] with a two-path notice naming both when it does not (case 3: an
archived worktree, so the operator learns where the pane landed instead of
finding out mid-edit). The selection logic is a pure function
(restore_member_cwd) so both cases are unit-tested without a real spawn.

Also: the tab-tree capture landed at 02d502b03 had a self-observation race.
attach()'s first push_layout(true) for a freshly-minted home squad now fires
before restore_squads runs a few lines later in the same call - eager
capture persisted that brand-new squad, and restore_squads then read its own
just-written row back as a "restored" lane matching its origin, home-merging
a duplicate pane into it. Caught by tests/persistence.rs
(persistence_zero_client_session_survives_and_resyncs_fully asserted 1 pane,
got 2) once the full suite ran, not by the scripted repro suite alone. Fixed
by making topology capture a no-op until self.restored flips true, which is
otherwise a one-shot gate for the server's lifetime.

Also: two pre-existing clippy warnings fixed in agents_view.rs (a redundant
u64->u64 cast on the macOS pid start-time read) while in the file.

+++ crates/fno/src/agents_view.rs
@@ -1107,7 +1107,7 @@ fn registry_start_time(pid: u32) -> Option<u64> {
     if written != size {
         return None;
     }
-    Some((info.pbi_start_tvsec as u64) * 1_000_000 + info.pbi_start_tvusec as u64)
+    Some(info.pbi_start_tvsec * 1_000_000 + info.pbi_start_tvusec)
 }
 
 #[cfg(target_os = "linux")]
